+++ pkg/util/rekor_client.go
@@ -1,0 +1,48 @@
+// Copyright 2021 The Sigstore Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package util
+
+import (
+	"net/url"
+
+	"github.com/go-openapi/runtime"
+	httptransport "github.com/go-openapi/runtime/client"
+	"github.com/go-openapi/strfmt"
+	"github.com/sigstore/rekor/pkg/generated/client"
+	"github.com/spf13/viper"
+)
+
+func GetRekorClient(rekorServerURL string) (*client.Rekor, error) {
+	url, err := url.Parse(rekorServerURL)
+	if err != nil {
+		return nil, err
+	}
+
+	rt := httptransport.New(url.Host, client.DefaultBasePath, []string{url.Scheme})
+	rt.Consumers["application/yaml"] = YamlConsumer()
+	rt.Consumers["application/x-pem-file"] = runtime.TextConsumer()
+	rt.Consumers["application/pem-certificate-chain"] = runtime.TextConsumer()
+	rt.Producers["application/yaml"] = YamlProducer()
+	rt.Producers["application/timestamp-query"] = runtime.ByteStreamProducer()
+	rt.Consumers["application/timestamp-reply"] = runtime.ByteStreamConsumer()
+
+	if viper.GetString("api-key") != "" {
+		rt.DefaultAuthentication = httptransport.APIKeyAuth("apiKey", "query", viper.GetString("api-key"))
+	}
+
+	registry := strfmt.Default
+	registry.Add("signedCheckpoint", &SignedCheckpoint{}, SignedCheckpointValidator)
+	return client.New(rt, registry), nil
+}
